folder structure, README.md
