+++ nlp-python/README.md
@@ -1,0 +1,9 @@
+### Run below commands in order to start python server
+
+    make install
+
+    make make auth 
+
+    make run
+
+#### Execute the APIs given in postman collection
