+++ nlp-python/main.py
@@ -3,20 +3,12 @@
 # Press ⌃R to execute it or replace it with your code.
 # Press Double ⇧ to search everywhere for classes, files, tool windows, actions, and settings.
 
-import vertexai
-from vertexai.language_models import TextGenerationModel
-from transformers import pipeline
-
-import os
-
 from flask import Flask, request
 
 from models import SentimentAnalysisEntity, TopicModellingEntity
+from nlp_utils import parts_of_speech, sentiment_analysis
 
 app = Flask(__name__)
-
-os.environ['GOOGLE_APPLICATION_CREDENTIALS'] = os.path.expanduser(
-    '~/.config/gcloud/application_default_credentials.json')
 
 
 @app.route('/health')
@@ -45,38 +37,6 @@ def sentiment_scoring():
     return {
         'sentimentScore': sentiment_analysis(sae.text)
     }, 200
-
-
-def parts_of_speech(tm_input):
-    vertexai.init(
-        project="ac-cntxtlbank-sct-project",
-        location="us-central1",
-    )
-    parameters = {
-        "temperature": 0.2,
-        "max_output_tokens": 256,
-        "top_p": 0.8,
-        "top_k": 40
-    }
-    model = TextGenerationModel.from_pretrained("text-bison@001")
-
-    response = model.predict(
-        tm_input,
-        **parameters
-    )
-    print(f"Response from Model: \n")
-    print(response.text)
-    return response.text
-
-
-def sentiment_analysis(input_to_transformer):
-    classifier = pipeline(model="distilbert-base-uncased-finetuned-sst-2-english")
-    result = classifier(input_to_transformer)
-    print(result)  # [{'label': 'POSITIVE', 'score': 0.9989916682243347}]
-    return {
-        'inputText': input_to_transformer,
-        'sentimentScore': result
-    }
 
 
 def print_hi(name):
